change Task38

+++ HomeWork5/Program.cs
@@ -39,10 +39,10 @@
 // Найдите разницу между максимальным и минимальным элементов массива.
 // [3 7 22 2 78] -> 76
 
-int [] array = {3, 7, 22, 2, 78};
-int diff = 0;
-int min= array[0];
-int max = array[0];
+double [] array = {3, 7, 22, 2, 78};
+double diff = 0;
+double min= array[0];
+double max = array[0];
 for (int i =1; i< array.Length; i++)
 {
 if (array[i] > max) max = array [i];
